**** General changes
Added performance test to the testing code of strhash.cc.

Added the ability to read the data from a file.

Improved the command line interface considerably by adding flags to
control whether performance and/or correctness tests are done and
whether data is randomly generated or read from a file and other
things.

Removed deleting the data after we are done as I didn't see the point
of it and it was made more subtle by the fact that the data is now in
a GrowArray.

Due to outstanding issues with hash function 2 given below, I changed
the default hash function back to hash function 1.

**** hash function 1 vs 2

Added a script to rebuild and test hash functions 1 and 2 versus each
other for speed: do_strhash_test

The following is written in a comment in strhash.cc:

We need to test them versus each other for randomness.  Scott suggests
simply hashing the same data and modding it down as a hashtable would
and seeing which gives more collisions.

Note that both hash functions could be improved if they were aware of
how many of their low order bits were really going to be preserved and
then taking the high order bits that would otherwise be discarded and
xoring or adding them to the low order bits.

**** hash function 2 changes

Renamed hash function 2 to the "word-rotate/final-mix" hash function.

Improved it by using a different large prime for each multiply and the
initial state.

The following Outstanding issues with hash function 2 which are noted
in the code in comments.

1) Does not work on architectures other than 32 bit.

2) Will not work well or at all if the strings do not start on 32-bit
aligned boundaries.

3) Could be made to work faster if the strings start AND end on 32-bit
aligned boundaries and are padded out with NUL-s.