remove 'override' macro, add standard 'override', add -std=c++11
sm_config.pm (bumped to version 1.07) now adds -std=c++11 to the
g++ command line by default, which is needed to accept the
'override' pseudo-keyword.

I have removed the 'override' macro (which, it appears, was only
used in one place in the entire verifier).

I manually inspected every instance of one class inheriting another,
adding the override keyword to all places that appeared to need it.