srcloc.cc: Run 'cpp' during test without "2>/dev/null" argument.
Under MinGW, 'system' invokes cmd.exe, which does not understand that
syntax for redirecting stderr.  Additionally, the error message when
the command fails shows the command line without that, meaning it was
added later (and the error message not updated).  I do not recall why
the stderr redirection was added, so I just discarded it.