Port smbase to GCC 6.4 with target x86_64-w64-mingw32.
One issue is this platform is compiled with GCC but uses the Microsoft C
run-time library, which is a combination I had not targeted before.

Furthermore, it has 32-bit 'long' but 64-bit pointers, necessitating the
use of stdint.h and uintptr_t; I had previously avoided those due to
portability concerns, but the passage of time has resolve that.