Add rectangle.h.
There is currently nothing that uses this file, either in smbase or
any client project.  I created this as part of an aborted attempt
to port the editor to FLTK, which lacks an equivalent of Qt's QRect.
Now, I want to keep this implementation, even though the immediate
need for it is gone, so I am tucking it away in smbase.