srcloc.cc: Remove "if (!this)" checks to silence GCC warnings.
Back in the day, which I was actively working on Elsa, there was a
certain common failure mode, and I added code to detect it and give a
nicer message.  But modern GCC elides the check since NULL 'this' is
undefined behavior, and tricking it into keeping the check is not worth
the hassle.

Removing this code also silences compiler warnings on it.  I had
previously silenced it with -Wno-nonnull-compare, but then that flag
itself was causing warnings on some versions of GCC, so now both the
code and the flag are gone, which is fine.