Rehabilitate and enable 'datablok' test.
I was surprised to learn it was not run with "make check".  I do not
remember why that might have been.

I had to fix the way memory corruption was being tested since it
previously only worked when NDEBUG was not set.  I also modified the
destructor to never throw an exception.