array.h: Check for allocating a negative length.
This arose from what I think is a memory corruption bug but it led to
trying to allocate a negative length, which then took me a minute or so
longer to realize than if I had had this assertion.  Since we are about
to do an allocation, the cost of the check is negligible.