refct-serf: Removed RCSerfC in favor of RCSerf<Foo const>.
After I wrote the rationale for having both RCSerf and RCSerfC, I
decided my reason was not very good, so went ahead and switched to
using a single class.  Not only is it less code, but it should be
more composable with other template classes since 'const' is part
of the template parameter rather than the name.