refct-serf: SerfRefCount no longer polymorphic.  Remove RCSerfBaseC.
This basically undoes my previous commit, then solves the multiple
inheritance problem in a different, better way.  The root of the
previous problem was the RCSerfBaseC class, which was the monomorphic
core of the parameteric-polymorphic (i.e., template) RCSerf.  But that
required storing a SerfRefCount* instead of a T*, leading to a variety
of problems and ugly solutions, including dynamic_cast.  The latter is
especially vexing since suddenly there is a possible performance reason
to avoid RCSerf, whereas I want to use it in as many places as I can.

Instead, I eliminated RCSerfBaseC in favor of two monomorphic helper
functions (packaged in class RCSerfPrivateHelpers) and a slightly
expanded RCSerf template implementation.

I also eliminated the ugly 'unsafe_getRCSerfBaseC' in favor of simply
having more flavors of 'failingSerfs' in the test program.