sm-file-util: getDirectoryEntries: Get file kind in one pass on Windows.
This makes it *much* faster to scan a directory to get all the files
and their kinds.  Like, from 700ms to 16ms for 100 scans.

The names-only scan is now called getDirectoryNames.

Add getFileKind to directly query file kind.  It and related functions
now throw xSysError on permission errors and similar.