Add on-leaving-scope.h.
This file was downloaded from
https://crascit.com/wp-content/uploads/2015/07/onleavingscope_update1.zip
which is linked from the blog whose URL is in the comments.
I then renamed it from OnLeavingScope.h to match the naming
conventions used elsewhere in smbase, but did not alter its
contents at all (yet).

license.txt: Acknowledge different license for on-leaving-scope.h,
and also mention that the provenance of crc.cpp has been lost to
the mists of time.