{bflatten,overflow}-test.cc: Wrap test code in anonymous namespace.
While testing with 'gcc -fsanitize=undefined', I saw:

  overflow.h:152:3: runtime error: load of value 2147483647, which is not a valid value for type 'SomeEnum'

which was perplexing because, in overflow-test.cc, 'SomeEnum' has
an enumerator with value 2147483647 (expressed as MAX_INT).  This
turned out to be caused by a collision with the 'SomeEnum' in
bflatten-test.cc (ODR violation), since the 'convertWithoutLoss'
template had already been instantiated with latter enum.

The fix is simple, using an anonymous namespace for test code, which
I'll do in the future instead of using 'static' on function
definitions.