nonport: Move tests into new nonport-test.cc.
The Makefile had a note saying not to link it with $(THIS), meaning
smbase.  That was because nonport is not supposed to depend on anything
else in smbase; it should be at the root of the dependency tree.  But
I no longer think that is an important property to ensure, and the
module is a poorly-designed grab-bag that I do not intend to add to or
use in new code, so there's no good reason to persist in treating it
specially.