syserr: Rename: m_context -> m_argument.
As noted in the now-deleted TODO, my original idea was to fold
m_context into m_contexts.  However, I realized that the former is
not really context at all, at least not in the same way that the
rest of m_context is, but instead pairs with the system call name.
So I decided to keep them separate and rename the field.